requirements.txt
# Core ML and Scientific Computing
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.6.0
gymnasium>=0.26.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.8.0
wandb>=0.12.0
scikit-learn>=1.1.0
scipy>=1.8.0
pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
# open3d>=0.15.0  # Optional: For point cloud processing

# Optional simulation dependencies (install separately)
# carla>=0.9.5  # Use: pip install carla==0.9.5
# msgpack-rpc-python>=0.4.1  # For AirSim
# msgpack>=1.0.0  # For AirSim

# Optional ROS dependencies (install separately)
# rospkg>=1.3.0
# sensor-msgs>=1.12.0
# geometry-msgs>=1.12.0
# nav-msgs>=1.12.0
# tf2-ros>=0.6.0
